Split mission.cpp and planner.cpp
