+++ orca_base/CMakeLists.txt
@@ -43,6 +43,7 @@ add_executable(
   base_node
   src/base_node.cpp
   src/mission.cpp
+  src/planner.cpp
   src/segment.cpp
   src/controller.cpp
   src/util.cpp
